[sprint 5][task 5] feat(auth-service): implement redis-backed 2fa code store
Added Redis-backed `TwoFactorAuthCodeStore` implementation for storing 2FA codes with a custom TTL. Updated configurations to include TTL for 2FA codes. Enhanced application startup logic to initialize the 2FA code store based on storage engine settings. Refactored and updated associated tests and documentation.

+++ auth-service/src/routes/login.rs
@@ -79,7 +79,7 @@ pub async fn login(
     let cookie = generate_auth_cookie(
         &user.email,
         config.jwt_secret.as_ref().unwrap(),
-        config.jwt_ttl_seconds,
+        i64::from(config.jwt_ttl_seconds),
     )?;
     let jar = jar.add(cookie);
     Ok((jar, StatusCode::OK.into_response()))
