+++ auth-service/src/routes/verify_2fa.rs
@@ -50,7 +50,7 @@ pub async fn verify_2fa(
     let cookie = generate_auth_cookie(
         &email,
         config.jwt_secret.as_ref().unwrap(),
-        config.jwt_ttl_seconds,
+        i64::from(config.jwt_ttl_seconds),
     )?;
     let jar = jar.add(cookie);
     Ok((jar, StatusCode::OK.into_response()))
